RavenDB-18537 removing old list view backup model
