+++ src/Raven.Studio/typescript/viewmodels/database/tasks/backups.ts
@@ -1,12 +1,7 @@
 import appUrl = require("common/appUrl");
 import router = require("plugins/router");
 import database = require("models/resources/database");
-import ongoingTasksCommand = require("commands/database/tasks/getOngoingTasksCommand");
-import ongoingTaskBackupListModel = require("models/database/tasks/ongoingTaskBackupListModel");
 import clusterTopologyManager = require("common/shell/clusterTopologyManager");
-import ongoingTaskModel = require("models/database/tasks/ongoingTaskModel");
-import deleteOngoingTaskCommand = require("commands/database/tasks/deleteOngoingTaskCommand");
-import toggleOngoingTaskCommand = require("commands/database/tasks/toggleOngoingTaskCommand");
 import databaseGroupGraph = require("models/database/dbGroup/databaseGroupGraph");
 import getDatabaseCommand = require("commands/resources/getDatabaseCommand");
 import ongoingTaskListModel = require("models/database/tasks/ongoingTaskListModel");
@@ -23,7 +18,6 @@ class backups extends shardViewModelBase {
     private clusterManager = clusterTopologyManager.default;
     myNodeTag = ko.observable<string>();
     
-    periodicBackupTasks = ko.observableArray<ongoingTaskBackupListModel>();
     recentManualBackup = ko.observable<manualBackupListModel>();
     
     isManualBackupInProgress = ko.observable<boolean>(false); // todo !!! create issue... for server ep
@@ -39,8 +33,6 @@ class backups extends shardViewModelBase {
     
     constructor(db: database) {
         super(db);
-        this.bindToCurrentInstance("confirmRemoveOngoingTask", "confirmEnableOngoingTask", "confirmDisableOngoingTask",
-                                   "toggleDetails", "createNewPeriodicBackupTask");
         this.initObservables();
     }
 
@@ -53,7 +45,7 @@ class backups extends shardViewModelBase {
 
     activate(args: any): JQueryPromise<any> {
         super.activate(args);
-        return $.when<any>(this.fetchDatabaseInfo(), this.fetchOngoingTasks(), this.fetchManualBackup());
+        return $.when<any>(this.fetchDatabaseInfo(), this.fetchManualBackup());
     }
 
     attached() {
@@ -98,7 +90,7 @@ class backups extends shardViewModelBase {
         if (!this.db) {
             return;
         }
-        return $.when<any>(this.fetchDatabaseInfo(), this.fetchOngoingTasks(), this.fetchManualBackup());
+        return $.when<any>(this.fetchDatabaseInfo(), this.fetchManualBackup());
     }
     
     private fetchDatabaseInfo() {
@@ -108,18 +100,7 @@ class backups extends shardViewModelBase {
                 this.graph.onDatabaseInfoChanged(dbInfo);
             });
     }
-
-    private fetchOngoingTasks(): JQueryPromise<Raven.Server.Web.System.OngoingTasksResult> {
-        const db = this.db;
-        return new ongoingTasksCommand(db, null) //TODO:
-            .execute()
-            .done((info) => {
-                info.OngoingTasksList = info.OngoingTasksList.filter(x => x.TaskType === "Backup");
-                this.processTasksResult(info);
-                this.graph.onTasksChanged(info);
-            });
-    }
-
+    
     private fetchManualBackup(): JQueryPromise<Raven.Client.Documents.Operations.Backups.GetPeriodicBackupStatusOperationResult> {
         const db = this.db.name;
         return new getManualBackupCommand(db)
@@ -133,144 +114,12 @@ class backups extends shardViewModelBase {
         this.fetchManualBackup();
     }
     
-    private watchBackupCompletion(task: ongoingTaskBackupListModel) {
-        if (!this.watchedBackups.has(task.taskId)) {
-            let intervalId = setInterval(() => {
-                task.refreshBackupInfo(false)
-                    .done(result => {
-                        if (!result.OnGoingBackup) {
-                            clearInterval(intervalId);
-                            intervalId = null;
-                            this.watchedBackups.delete(task.taskId);
-                        }
-                    })
-            }, 3000);
-
-            this.watchedBackups.set(task.taskId, intervalId as unknown as number);
-
-            this.registerDisposable({
-                dispose: () => {
-                    if (intervalId) {
-                        clearInterval(intervalId);
-                        intervalId = null;
-                        this.watchedBackups.delete(task.taskId);
-                    }
-                }
-            });
-        }
-    }
-    
     toggleDetails(item: ongoingTaskListModel) {
         item.toggleDetails();
     }
     
     private processManualBackupResult(dto: Raven.Client.Documents.Operations.Backups.GetPeriodicBackupStatusOperationResult) {
        this.recentManualBackup(dto.Status ? new manualBackupListModel(dto.Status) : null);
-    }
-    
-    private processTasksResult(result: Raven.Server.Web.System.OngoingTasksResult) {
-        const oldTasks: Array<{ taskId: number }> = [
-            ...this.periodicBackupTasks()
-        ];
-
-        const oldTaskIds = oldTasks.map(x => x.taskId);
-        const newTaskIds = result.OngoingTasksList.map(x => x.TaskId);
-        const toDeleteIds = _.without(oldTaskIds, ...newTaskIds);
-
-        const groupedTasks = _.groupBy(result.OngoingTasksList, x => x.TaskType);
-       
-        this.mergeTasks(this.periodicBackupTasks, 
-            groupedTasks["Backup" as Raven.Client.Documents.Operations.OngoingTasks.OngoingTaskType], 
-            toDeleteIds,
-            (dto: Raven.Client.Documents.Operations.OngoingTasks.OngoingTaskBackup) => new ongoingTaskBackupListModel(dto, task => this.watchBackupCompletion(task)));
-        
-        // Sort backup tasks 
-        const groupedBackupTasks = _.groupBy(this.periodicBackupTasks(), x => x.isServerWide());
-        const serverWideBackupTasks = groupedBackupTasks.true;
-        const ongoingBackupTasks = groupedBackupTasks.false;
-
-        if (ongoingBackupTasks) {
-            this.periodicBackupTasks(serverWideBackupTasks ? ongoingBackupTasks.concat(serverWideBackupTasks) : ongoingBackupTasks);
-        } else if (serverWideBackupTasks) {
-            this.periodicBackupTasks(serverWideBackupTasks);
-        }
-    }
-    
-    private mergeTasks<T extends ongoingTaskListModel>(container: KnockoutObservableArray<T>, 
-                                                       incomingData: Array<Raven.Client.Documents.Operations.OngoingTasks.OngoingTask>, 
-                                                       toDelete: Array<number>,
-                                                       ctr: (dto: Raven.Client.Documents.Operations.OngoingTasks.OngoingTask) => T) {
-        // remove old tasks
-        container()
-            .filter(x => _.includes(toDelete, x.taskId))
-            .forEach(task => container.remove(task));
-        
-        (incomingData || []).forEach(item => {
-            const existingItem = container().find(x => x.taskId === item.TaskId);
-            if (existingItem) {
-                existingItem.update(item);
-            } else {
-                const newItem = ctr(item);
-                const insertIdx = _.sortedIndexBy(container(), newItem, x => x.taskName().toLocaleLowerCase());
-                container.splice(insertIdx, 0, newItem);
-            }
-        });
-    }
-
-    confirmEnableOngoingTask(model: ongoingTaskModel) {
-        const db = this.db;
-
-        this.confirmationMessage("Enable Task", "You're enabling task of type: " + model.taskType(), {
-            buttons: ["Cancel", "Enable"]
-        })
-            .done(result => {
-                if (result.can) {
-                    new toggleOngoingTaskCommand(db, model.taskType(), model.taskId, model.taskName(), false)
-                        .execute()
-                        .done(() => model.taskState('Enabled'))
-                        .always(() => this.fetchOngoingTasks());
-                }
-            });
-    }
-
-    confirmDisableOngoingTask(model: ongoingTaskModel) {
-        const db = this.db;
-
-        this.confirmationMessage("Disable Task", "You're disabling task of type: " + model.taskType(), {
-            buttons: ["Cancel", "Disable"]
-        })
-            .done(result => {
-                if (result.can) {
-                    new toggleOngoingTaskCommand(db, model.taskType(), model.taskId, model.taskName(), true)
-                        .execute()
-                        .done(() => model.taskState('Disabled'))
-                        .always(() => this.fetchOngoingTasks());
-                }
-            });
-    }
-
-    confirmRemoveOngoingTask(model: ongoingTaskModel) {
-        const db = this.db;
-        
-        this.confirmationMessage("Delete Task", "You're deleting task of type: " + model.taskType(), {
-            buttons: ["Cancel", "Delete"]
-        })
-            .done(result => {
-                if (result.can) {
-                    this.deleteOngoingTask(db, model);
-                }
-            });
-    }
-
-    private deleteOngoingTask(db: database, model: ongoingTaskModel) {
-        new deleteOngoingTaskCommand(db, model.taskType(), model.taskId, model.taskName())
-            .execute()
-            .done(() => this.fetchOngoingTasks());
-    }
-    
-    createNewPeriodicBackupTask() {
-        const url = appUrl.forEditPeriodicBackupTask(this.db);
-        router.navigate(url); 
     }
 
     createManualBackup() {
